Splitting the update fields into seperate components
